+++ test/providers/livechat.js
@@ -14,11 +14,10 @@ describe('LiveChat', function () {
       expect(window.__lc).not.to.be(undefined);
       expect(window.LC_API).to.be(undefined);
 
-      // When the library loads, `LC_API` and `LC_Invite` will be defined.
+      // When the library loads, `LC_API` will be defined.
       var interval = setInterval(function () {
-        if (!window.LC_API || !window.LC_Invite) return;
+        if (!window.LC_API) return;
         expect(window.LC_API).not.to.be(undefined);
-        expect(window.LC_Invite).not.to.be(undefined);
         expect(spy.called).to.be(true);
         clearInterval(interval);
         done();
@@ -34,31 +33,30 @@ describe('LiveChat', function () {
 
 
   describe('identify', function () {
-
-    beforeEach(analytics.user.clear);
+    var stub;
+    beforeEach(function () {
+      stub = sinon.stub(window.LC_API, 'set_custom_variables');
+      analytics.user.clear();
+    });
+    afterEach(function () { stub.restore(); });
 
     it('should set user id', function () {
-      var stub = sinon.stub(window.LC_API, 'set_custom_variables');
       analytics.identify(test.userId);
       expect(stub.args[0][0]).to.eql([
         { name : 'User ID', value : test.userId }
       ]);
-      stub.restore();
     });
 
     it('should set traits', function () {
-      var stub = sinon.stub(window.LC_API, 'set_custom_variables');
       analytics.identify(test.traits);
       expect(stub.args[0][0]).to.eql([
         { name : 'name', value : test.traits.name },
         { name : 'email', value : test.traits.email },
         { name : 'created', value : test.traits.created }
       ]);
-      stub.restore();
     });
 
     it('should set user id and traits', function () {
-      var stub = sinon.stub(window.LC_API, 'set_custom_variables');
       analytics.identify(test.userId, test.traits);
       expect(stub.args[0][0]).to.eql([
         { name : 'User ID', value : test.userId },
@@ -66,9 +64,6 @@ describe('LiveChat', function () {
         { name : 'email', value : test.traits.email },
         { name : 'created', value : test.traits.created }
       ]);
-      stub.restore();
     });
-
   });
-
 });