+++ test/providers/olark.js
@@ -33,41 +33,44 @@ describe('Olark', function () {
 
   describe('identify', function () {
 
-    beforeEach(analytics.user.clear);
+    var spy;
+    beforeEach(function () {
+      analytics.user.clear();
+      spy = sinon.spy(window, 'olark');
+    });
+    afterEach(function () { spy.restore(); });
 
-    it('should updates visitor nickname with the best name', function () {
-      var spy = sinon.spy(window, 'olark');
+    it('should update with the email', function () {
       analytics.identify({
         email : 'zeus@segment.io'
       });
-
       expect(spy.calledWithMatch('api.chat.updateVisitorNickname', {
         snippet : 'zeus@segment.io'
       })).to.be(true);
+    });
 
-      spy.reset();
+    it('should update with the name and email when called', function () {
       analytics.identify(test.traits);
       expect(spy.calledWithMatch('api.chat.updateVisitorNickname', {
         snippet : 'Zeus (zeus@segment.io)'
       })).to.be(true);
+    });
 
-      spy.reset();
+    it('should update with the user id', function () {
       analytics.identify(test.userId);
       expect(spy.calledWithMatch('api.chat.updateVisitorNickname', {
         snippet : test.userId
       })).to.be(true);
+    });
 
-      spy.reset();
+    it('should update with the email, name, and traits with userId', function () {
       analytics.identify(test.userId, test.traits);
       expect(spy.calledWith('api.chat.updateVisitorNickname', {
         snippet : 'Zeus (zeus@segment.io)'
       })).to.be(true);
-
-      spy.restore();
     });
 
     it('should update the visitor email', function () {
-      var spy = sinon.spy(window, 'olark');
       analytics.identify({
         email : 'zeus@segment.io'
       });
@@ -75,12 +78,9 @@ describe('Olark', function () {
       expect(spy.calledWithMatch('api.visitor.updateEmailAddress', {
         emailAddress : 'zeus@segment.io'
       })).to.be(true);
-
-      spy.restore();
     });
 
     it('should update the visitor phone number', function () {
-      var spy = sinon.spy(window, 'olark');
       analytics.identify({
         phone : '(555) 555-5555'
       });
@@ -88,12 +88,9 @@ describe('Olark', function () {
       expect(spy.calledWithMatch('api.visitor.updatePhoneNumber', {
         phoneNumber : '(555) 555-5555'
       })).to.be(true);
-
-      spy.restore();
     });
 
     it('should update the visitor full name', function () {
-      var spy = sinon.spy(window, 'olark');
       analytics.identify({
         name : 'Hallucinating Chipmunk'
       });
@@ -109,27 +106,17 @@ describe('Olark', function () {
       expect(spy.calledWithMatch('api.visitor.updateFullName', {
         fullName : 'Hallucinating Chipmunk'
       })).to.be(true);
-
-      spy.restore();
     });
 
     it('should update the visitor custom fields', function () {
-      var spy = sinon.spy(window, 'olark');
       analytics.identify({
         dogs : 1
       });
       expect(spy.calledWithMatch('api.visitor.updateCustomFields', {
         dogs : 1
       })).to.be(true);
-
-      spy.restore();
     });
-
-
-          
-
   });
-
 
   describe('track', function () {