+++ test/utils.js
@@ -40,6 +40,11 @@ window.test = {
   // Providers
   // ---------
 
+  'AdRoll' : {
+    advId : 'LYFRCUIPPZCCTOBGRH7G32',
+    pixId : 'V7TLXL5WWBA5NOU5MOJQW4'
+  },
+
   'Bitdeli' : {
     inputId   : 'x',
     authToken : 'y'
@@ -115,6 +120,8 @@ window.test = {
   'Olark' : '5798-949-10-1692',
 
   'Perfect Audience' : '4ff6ade4361ed500020000a5',
+
+  'Pingdom' : '5168f8c6abe53db732000000',
 
   'Quantcast' : 'x',